# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Deleted": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "Deleted",
        "type": "object"
      },
      "DocList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DocList",
        "type": "object"
      },
...
      "DocumentCreate": {
        "properties": {
          "background": {
            "default": "transparent",
            "title": "Background",
            "type": "string"
          },
          "from_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Asset"
          },
          "h": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "H"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "w": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "W"
          }
        },
        "title": "DocumentCreate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "loom2 orchestrator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/documents": {
      "get": {
        "operationId": "documents_list_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Documents List",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "documents_create_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Create",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}": {
      "delete": {
        "operationId": "documents_delete_documents__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Delete",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "documents_get_documents__doc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Get",
        "tags": [
          "documents"
        ]
      },
      "put": {
        "description": "Replace the stack. The reply lists raster layers / masks the server has no bytes for, so the editor uploads\nthem before `/save` even when it does not consider them dirty (B15: delete · save · undo · save).",
        "operationId": "documents_put_documents__doc_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPutReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Put",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/ai": {
      "post": {
        "description": "Queue an inpaint / refine / upscale job on this document (10 §13); results come back as layers\n(`document.changed` events with `added`) or, for upscale, as a Catalogue asset too.",
        "operationId": "documents_ai_documents__doc_id__ai_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsSubmitted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Ai",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/close": {
      "post": {
        "operationId": "documents_close_documents__doc_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Closed"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Close",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/compare": {
      "post": {
        "description": "Raw straight-alpha RGBA of the editor's GPU composite (w·h·4 bytes) → per-channel delta against the exact\nflatten (10 §14 item 1). Both images are kept under temp/compare for inspection; the result lands in doc.meta.",
        "operationId": "documents_compare_documents__doc_id__compare_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "w",
            "required": true,
            "schema": {
              "title": "W",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "h",
            "required": true,
            "schema": {
              "title": "H",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Compare",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/export": {
      "post": {
        "operationId": "documents_export_documents__doc_id__export_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Export",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/flatten": {
      "post": {
        "operationId": "documents_flatten_documents__doc_id__flatten_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlattenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FlattenedToCatalogue"
                    },
                    {
                      "$ref": "#/components/schemas/FlattenedToFile"
                    }
                  ],
                  "title": "Response Documents Flatten Documents  Doc Id  Flatten Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Flatten",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/layers/{lid}/pixels": {
      "get": {
        "operationId": "layer_pixels_get_documents__doc_id__layers__lid__pixels_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lid",
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "image",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "raw",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Raw",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Layer Pixels Get",
        "tags": [
          "documents"
        ]
      },
      "put": {
        "description": "Raw uint8 body: RGBA (w·h·4 bytes) for images, grey (w·h) for masks; the fast path from the compositor (E2).",
        "operationId": "layer_pixels_put_documents__doc_id__layers__lid__pixels_put",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "lid",
            "required": true,
            "schema": {
              "title": "Lid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "w",
            "required": true,
            "schema": {
              "title": "W",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "h",
            "required": true,
            "schema": {
              "title": "H",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "image",
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PixelsPut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Layer Pixels Put",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/save": {
      "post": {
        "operationId": "documents_save_documents__doc_id__save_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Save",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/selection": {
      "get": {
        "description": "The document's selection as raw grey bytes (the AI Select result lands here, M5 slice 2).",
        "operationId": "documents_selection_get_documents__doc_id__selection_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Selection Get",
        "tags": [
          "documents"
        ]
      },
      "put": {
        "description": "The editor's selection as raw grey bytes (w·h); an empty body clears it. The M5 edit jobs read it.",
        "operationId": "documents_selection_put_documents__doc_id__selection_put",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "w",
            "required": false,
            "schema": {
              "default": 0,
              "title": "W",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "h",
            "required": false,
            "schema": {
              "default": 0,
              "title": "H",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionReply"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Selection Put",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{doc_id}/thumbnail": {
      "get": {
        "operationId": "documents_thumbnail_documents__doc_id__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Documents Thumbnail",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}